Remove stale binary previews from output ODF file
The ODF original file includes a binary graphic version of all of the
graphs.  These graphs are no longer valid once we've replaced the data
in the sheet, so remove them from the file and manifest to ensure
OpenOffice will rebuild the graph using our new data.  Same for the
thumbnail image.