Fix startup to work in paths with spaces
The EZFIO.BAT script needs to invoke a couple levels of PowerShell
to start our script, and if quoting rules are not carefully followed
this will fail if there is a space anywhere in the path to the script.
Rewrote and tested the bouncer batch file using a directory including
a space.  Fixes issue 14 and issue 8.