Add exceedance charting to main branch
Merge in the changes in the exceedance branch, as wall as additional
support for the latest FIO 2.99 which has a different latency report
format (lat_ns, no weird semi-log implied latencies).

Windows exceedance generation not tested w/2.99 since the latest
bluestop release is only 2.18, but that version does work w/the PS1.
At some point when new release hits Windows we'll need to do the same
changes as in Linux.