Preliminary stab at combining utility
Takes two complete ODS runs and combine them into a single, merged
file.

This seems to work in preliminary testing, but needs more work to
be robust and ensure correctness.  It's doing a lot of ugly XML
hacking using regular expressions, which is not the method most
people would use...with good cause.