Add --fast-precondition to skip random fill stage
On HDDs or massive SSD arrays, writing TB of data with 4K random chunks
can literally take days or weeks.  Allow for skipping the random
preconditioning tests by a new --fast-precondition flag.

The test results may be different than the standard, full precondition
mode, so the generated ODS will have a prefix in the ODS cells that
describe the SUT.

Only 1 sequential fill is performed, to ensure all user-accessible
blocks are allocated in this mode.