Fix key_error[bins] on newer FIO versions
FIO head now omits the bins[] array for test results when there are no entries
in the actual test (i.e. on a pure read test, the write key won't have a "bins").
This could cause a key_error on running the first test of the pattern.

Now, check if there is a "N" key, and if so that it is greater than 0 to avoid
looking for a 'bins' that isn't there.

Works on latest FIO head and older versions.