Add --readonly flag to not disturb existing data
In some cases it makes sense to test performance of a device after it has been
filled by a real applicaiton and filesystem.  Add a "--readonly" flag which
skips any tests that have non-0 write percentages (and, of course, the
preconditioning stages).  The results aren't really comparable to a normal run,
but can still be useful.